(#237)refactor: 공지 service interface 구현했습니다.
